apps/requirements.txt
fastapi==0.100.0
uvicorn==0.23.2
sqlalchemy[asyncio]==2.0.15
asyncpg==0.27.0
pydantic==2.1.1
alembic==1.11.2
asyncpg==0.27.0
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
python-jose==3.3.0
pytest
httpx==0.27.0
pytest-asyncio==0.25.3
transformers==4.49.0
faiss-cpu==1.10.0
chromadb==0.6.3
sentence-transformers==3.4.1
pydantic[email]
python-multipart